move killswitch tests, just wait
